feat(nix/buildLisp): implementation specific deps and srcs
Both the deps and srcs arguments may now instead of their normal
contents receive filter sets of the following form:

  {
    impl1 = …;
    impl2 = …;
    …
    default = …; # optional
  }

When evaluating a buildLisp derivation these filters are processed and
will be applied. This is useful to express implementation specific
inclusion of (extra) files, implementation specific (extra) dependencies
on bundled and non-bundled dependencies.

Of course this is somewhat limited and can't have the power of an asdf
system definition in this regard (which is probably a good thing). I'm
quite happy with the usability and expressiveness of this so far,
however it significantly complicates the implementation of buildLisp and
I'm not yet confident this is bug-free (which is always a scary thought
with Nix). An alternative, possibly simpler solution would be to allow
buildLisp to receive a function as its argument which is then called
with the name of the actually used implementation and leave it to
downstream consumers to do the special casing.

Change-Id: I082c3701d1f5063b92100bf336a83425471c269d