refactor(pipelines/depot): prevent .drv files from being gc'd
Instead of falling back to reevaluating if the .drv file is gc'd, I
discovered while reading the Nix documentation [1] that we can
actually use a { path = true; } string context to create a reference to
a .drv file, so it can't be gc'd as long as the depot.yaml symlink of
the static pipeline exists (I'm not quite sure when this symlink is
deleted though).

[1]: Obviously I was joking, the string context builtins are completely
undocumented. I found the answer to my question in this test case:
https://github.com/NixOS/nix/blob/master/tests/lang/eval-okay-context-introspection.nix

Change-Id: Ia95565296921263411b58ccc58457b510cde0c11