refactor(pipelines/depot): prevent .drv files from being gc'd
Instead of falling back to reevaluating if the .drv file is gc'd, I
discovered while digging through the Nix source code [1] that we can
actually use a { path = true; } string context to create a reference to
a .drv file, so it can't be gc'd as long as the depot.yaml symlink of
the static pipeline exists. Thus we create a temporary gcroot for the
current pipeline run (uniquified by BUILDKITE_BUILD_ID) which is then
cleaned up in :duck: (even in the fallback pipeline in case
buildkite-agent upload fails). To avoid copying the same path to a
thousand places, we add a Buildkite pre-command hook [2] which is a bit
annoying, but much less cumbersome than using the meta-data mechanism to
share a (static) value between the steps.

[1]: See e. g. https://github.com/NixOS/nix/blob/master/tests/lang/eval-okay-context-introspection.nix
[2]: https://buildkite.com/docs/agent/v3/hooks

Change-Id: Ia95565296921263411b58ccc58457b510cde0c11