feat: executable nix test suites using nint
How nix test suites currently is more or less flawed:

* Despite the test suites essentially boiling down to a `true` or
  `false` result _at evaluation time_, we currently transform this
  boolean into noop derivations which either succeed or fail.

  This a) litters the nix store and b) doesn't parallelize well.

* Since cl/3427 we filter out already realized derivations from the CI
  output. This means that if something changes which tested by the test
  suite, the test suite will only show up in the CI output if the change
  causes it to fail. This is confusing for users. We can detect test
  suite changes to a certain extent (cl/3536), but won't ever be able to
  detect dependency changes.

To work around this, we construct test suites in such a way that they
are executable using //nix/nint and readTree can still import them.
The test suite will then be executed by nint as an (uncoditional) CI
target and not at eval time.

Additionally we get nicer test suite output in both failure and success
cases. One downside is that this adds more static CI target, since it is
still not easy to detect if a test suite needs to be re-run or not.

All easy to convert test suites now use the new scheme. The remaining
ones don't use readTree for the test suite which wouldn't be detected in
CI constructions.

Change-Id: Ib7a476d7f31caf6dd56b54f82bf3578f32437091