feat(web/tvl): Check in first version of new TVL logo
The exported SVG was hand-edited to make it as understandable as
possible, the components within it are grouped sensibly.

Note that the inside of the "armchair" is currently transparent, but
we should figure out a way to apply a fill to within the boundaries of
that paths (extra complicated because the "top" of the armchair is
actually the lower body of the virus, which is in a separate SVG
group).

Thanks Varia <3

Change-Id: Ifdb5f4ff7827caf10d193b99e81b7c8498b35ce4