feat(ops/pipelines): persist drvmap structure for each commit
Always create a structure that maps all targets to derivations, and
persist it as a JSON file.

This relates to some of the ideas expressed in:

https://docs.google.com/document/d/16A0a5oUxH1VoiSM8hyFyLW0WiUYpNo2e2D6FTW4BlH8/edit

The file is always uploaded to Buildkite as an artifact where it will
be stored for a number of days.

For builds on canon, the structure is stored as a git note in
refs/notes/target

Change-Id: I6d808461cd1c1fdd6983ba8c8ef075736d42caa7