chore(wpcarro): Simplify README.md
I rewrote my `README.md` most recently to show some managers what type
of side-projects I was working on. After successfully transfering to SRE
internally, I don't need the `README.md` to be a marketing document but
rather a tutorial for my future self. This change is a step in that
direction.

Change-Id: Ieaf0e72c8a33a163e6b6adefd76665ca675e8462