chore(depot): Configure refs/rev/* in Gerrit
This is a set of refs which can be pushed by CI and which lets us refer
to submitted changes in order.

For example, cl/3780 might be merged and become rev/3000. cl/3875 might
be merged later and become rev/3001, despite being an older review.

This is advantageous for things like deployments where it is more
interesting to refer to things in submit order than review order.

Note that this is different from the strategy employed at some companies
where reviews and submitted changes are numbered in the same number
sequence, but this is difficult to accomplish with Gerrit in a smooth
way.

Change-Id: I2e3df371e63a6a1836b443498a304f7526360a73