fix(third_party/nix-diff): use version with better decode errors
We had a problem on whitby where decoding of the drv files would fail
with an utf8-decoding error. My best guess is that it’s trying to read
a binary file, but since the error had no paths in it it was hard to
confirm.

This version of nix-diff will throw a better error message in that
case:
https://github.com/Profpatsch/nix-diff/commit/ee5bd082d2d780b7ff0a82ae7ae6dd73aafb68be

Change-Id: I5cb245923c6db0875e63e420cb0783e235b6859f