fix(depot): Disable topic change permissions
The reason for this is a bit annoying: besadii is moving to different
Gerrit hooks than before (cl/3928), as part of these hooks Gerrit passes
a `--topic` flag which contains the string value of the topic of the CL.

However, if no topic is set, Gerrit passes this flag *anyways* which
complicates arg parsing as Go's flag parsing has no way to ignore
unknown flags and there is no known type for the flag (it either has no
argument, i.e. acts like a boolean flag, or it has a string argument).

There are a few options to get out of this:

* use a 3rd-party flag parser in besadii
* rewrite the flag parsing ourselves
* rewrite besadii in something else
* ban topic changes

For now, to get this working and since we don't really use them, I'm
opting for the last option here.

Change-Id: I0fc3108dacf3a148776b5dc6d47d9c5f99b4d8de