feat(rules.pl): Allow subtree commits to be merged
Adds a Gerrit 'submit_type' rule which determines the type of submit
for a given commit.

By default, all commits are rebased (if necessary), leading to a
linear repo history.

The exception is a new type of commit - subtree commits - which
instead receive 'Merge if necessary' treatment.

This means that we can do subtree operations without having to change
Gerrit settings, simply by staging them on `subtree-staging`.

Change-Id: Iee6c616ff690d5f585968b8924e9bc483c354471