feat(ops/besadii): Strip patchset from branch in builds
The branch key for buildkite builds isn't actually used to fetch if a
commit is given - instead, it's just a visual grouping of multiple
builds. If we strip the patchset number from the branch, then, we can
get visual grouping of builds of successive patchsets of the same CL
number, even though the ref we're providing isn't a real ref.

Change-Id: Iaa9111297a88f965fda94cd8266240106f58a100