fix(grfn/emacs): More robust org-element title extraction
It *appears* that the main trigger for the :title of an org element
being a list instead of a string is if it contains rich text properties
as children, eg links, bold text, etc. This defines a function
that *attempts* to turn all of those into a single string, for use in my
i3status bar for the currently clocked-in task.

Change-Id: Ib655595914e54cd0e01f0fee6d1352c98a9f1436