fix(panettone): Fix export of issue status to SQL
Postmodern changed how users customize the way lisp values get exported
to SQL - now, in addition to defining methods of
`cl-postgres:to-sql-string`, we have to pass `:col-export` and
`:col-import` args to the field itself in the dao class.

I'm not *entirely* sure why both are necessary, but without both this
doesn't work.

Change-Id: Iae8fb63c34fb6c79b9dfa350129032aab5cd2233