fix(3p/overlays): Attempt to fix hard-to-debug suspend issue
I have an issue where my AMD laptop occasionally does not wake from
suspend. The only thing ever indicated in the logs is related to the
patched line here:

  xss-lock[2409537]: saver_blank: line 19: exec: sleep: not found

This stems from xsecurelock calling a script that in turn calls sleep,
in a loop.

I don't know if this causes my unsuspend problem because this also
happens sometimes without anything failing, but it's the best lead
I've got for now. Will run with this fix for a while and see if
anything changes in the stability of suspend.

Change-Id: Ie8190433c2125cbeddb909013be9dfa00141d5c6