fix(web/panettone): Only build one copy
Previously, due to the buildLisp implementation machinery, we'd have a
target for panettone, plus a target for panettone:sbcl. This was
relatively harmless (if a little noisy) before, but now that we have
integration tests is resulting in us running two copies of the
integration test step, which is no good. To work around that, this
commit makes buildLisp not override passthru.meta.targets set on its
arguments, and then passes an empty list to targets for panettone so we
only have one copy.

Change-Id: I1b624d58eb093a010a0be1f207e393252487a729