feat(rules.pl): Determine most applicable change owner on autosubmit
When autosubmitting changes, clbot may rebase a change on canon before
submitting. This action makes `clbot` both the uploader of the change
and the committer of the rebased commit.

For the purposes of our checks, this caused a few weirdnesses:

- All-Comments-Resolved and Conformant-Commit-Message were marked as
  having been resolved by clbot
- If owners permissions of the uploader were used to greenlight the
  change, they would get lost since clbot itself is not a code owner

This change instead introduces a new Prolog fact `applicable_owner`
which determines the most applicable user to consider the owner of a
change. Unless the uploader is clbot, this is always the uploader.

If the uploader is clbot, then the change owner is used instead.

This means:

  * patchsets pushed to a change by a non-owning user DO NOT get CRFO
  * patchsets rebased by clbot for autosubmit DO iff approvals were
    sufficient for submission at the time of the rebase

It also introduces an unusual edge case: If a patchset was uploaded by
somebody other than the owner of the change, and got CRFO from either
the new uploader or a third person adding a code review, the rebased
patchset _may_ instead gain CRFO from the change owner.

I think this may introduce an obscure vulnerability where somebody
could get a patchset to submit via someone else's CL, but only if
either another reviewer approved the change already or the fraudster
has these permissions themselves.

To guard against abuse of this a subsequent change will tighten the
permissions on copying Code-Review scores.

Change-Id: I5ac166650a3e21f8579994549f1dd19dff441572