feat(project.config): Only copy approval on identical file lists
This copies the definition of the Code-Review label from the
All-Projects project in Gerrit so that we can override parts of it.
Gerrit does not allow merging of label configurations across the
hierarchy, so the whole definition is placed here.

Specifically, instead of setting copyAllScoresOnTrivialRebase we now
set copyAllScoresIfListOfFilesDidNotChange and *no longer* set
`copyValue = +2`.

This means that review approvals are ONLY copied forward to new
patchsets if no new files were added.

Reasoning:

- guards against cases where a user tricks somebody with approval
  permission for their attack target into approving a benign CL, and
  then replacing the file list

- cl/4506 introduces a potential circumvention of approval rules in
  some limited circumstances which are substantially mitigated here

Submit-blocking reviews (-2) are always copied forward.

Change-Id: I245eea85ce5d214e18163b1fb8105661e7300d27
Reviewed-on: https://cl.tvl.fyi/c/depot/+/4507
Tested-by: tazjin <mail@tazj.in>
Reviewed-by: sterni <sternenseemann@systemli.org>
Reviewed-by: grfn <grfn@gws.fyi>