feat(wpcarro/marcus): Require python3
I use `python3` REPL as a calculator most often.

Side note: I don't have a clear model of which binaries belong in NixOS
configurations and which belong in my Emacs configuration. Here's a sketch of
what I'm thinking:

- shared (wpcarro's coreutils for fzf, ripgrep, git, etc)
  - marcus
  - diogenes
  - wpcarros-emacs (see footnote)

Footnote: It might be redundant to install wpcarro's coreutils in Emacs on NixOS
environments, but I install wpcarros-emacs in non-NixOS environments, so I
depend on them there.

Change-Id: Ib2b87c8e4e04d167ccbf837e0f85606cbf272828