refactor(web/todolist): use jq instead of Nix for TODO generation
I suppose the jq variant has the benefit that it supports text
escaping (which could be ported to the extract script if necessary),
in the rare case of rust updates or such it doesn't slow down
the pipeline evaluation, and it fixes a few bugs in the output HTML
(namely invalid paragraph (<p>) tags).

original motivation:

although it is nice to be able to generate something like this
using Nix, `fromJSON (readFile ...)` slows down the pipeline
evaluation a bit (~1 second) and can't be done in parallel to
other build jobs. This CL "outsources" this generation to jq
(especially because we don't gain much from using Nix here),
which results in the benefit that we don't have to wait for
the results of ripgrep+jq+readFile+fromJSON, which instead
just gets put into the resulting pipeline, but does not block
pipeline generation itself.

Change-Id: I5baeecc4ccd8c134675616e73ace5d898c452955